requirements.txt
gradio==4.19.2
pydub==0.25.1
google-generativeai==0.3.2
openai==1.5.0
pysrt==1.1.2
python-dotenv==1.0.0
requests==2.31.0